+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 - Removed RcVecUtils.Dot()
 - Removed RcVecUtils.Scale()
+- Removed RcVecUtils.Subtract(RcVec3f i, float[] verts, int j)
 
 ### Special Thanks
 - [@Doprez](https://github.com/Doprez)